crud provider

+++ src/main/java/cmcglobal/ebook/service/impl/ProviderService.java
@@ -3,6 +3,7 @@ package cmcglobal.ebook.service.impl;
 import cmcglobal.ebook.entity.Book;
 import cmcglobal.ebook.entity.Provider;
 import cmcglobal.ebook.exception.ExceptionHandle;
+import cmcglobal.ebook.repository.IBookRepository;
 import cmcglobal.ebook.repository.IProviderRepository;
 import cmcglobal.ebook.service.IBookService;
 import cmcglobal.ebook.service.IService;
@@ -19,6 +20,9 @@ public class ProviderService implements IService <Provider> {
     @Autowired
     IProviderRepository providerRepository;
 
+    @Autowired
+    IBookRepository bookRepository;
+
     @Override
     public List<Provider> findAll() {
         return providerRepository.findAll();
@@ -26,12 +30,14 @@ public class ProviderService implements IService <Provider> {
 
     @Override
     public Provider findById(Long id) {
-        return providerRepository.getById(id);
+        return providerRepository.getProviderById(id);
+//        Provider provider = providerRepository.getById(id);
+//        return provider;
     }
 
     @Override
-    public Provider findByName(String name) {
-        return providerRepository.getProviderByName(name);
+    public Provider findByCode(String name) {
+        return providerRepository.getProviderByCode(name);
     }
 
     @Override
@@ -39,10 +45,14 @@ public class ProviderService implements IService <Provider> {
         providerRepository.save(elemenInput);
     }
 
+
+
+
     @Override
     public void delete(Long id) {
         Provider provider = providerRepository.getById(id);
-        if(provider !=null){
+        Book book = bookRepository.getBookByProvider(provider);
+        if(provider !=null && book==null){
             providerRepository.delete(provider);
         }
